fix: Move statsd package under api and impl
This fixes GraalVM smoke test, but I would think the issue could have broken
regular agent.
