+++ dd-java-agent/instrumentation-testing/src/main/java/datadog/trace/agent/test/BootstrapClasspathSetupListener.java
@@ -69,7 +69,6 @@ public class BootstrapClasspathSetupListener implements LauncherSessionListener 
     "datadog.instrument",
     "datadog.appsec.api",
     "datadog.metrics.api",
-    "datadog.metrics.statsd",
     "datadog.trace.api",
     "datadog.trace.bootstrap",
     "datadog.trace.config.inversion",
